chore: Abstract application configuration
Create the `AppConfig` type to abstract the application configuration.

Create the `StartApp()` function in api/startg.go to abstract the
application starting.

Add the godotenv library and use it to load the environment variables
to set their to the `AppConfig` value.

Add the support for the `APP_ENVIRONMENT` environment configuration
parameter, with the possible values so far being:

- Debugging
- Development (Default - if none is specified).
- Testing.
- Production.

Change the .env file so that the application can run with a
`go run cmd/main.go` in the host system, without being in a container.